ai_service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.100.0
python-multipart>=0.0.6
uvicorn>=0.23.0
python-dotenv>=1.0.0
